chore(adsp-feedback): adjusting styles
Also moving type definitions in declaration file as reference.
