+++ libs/adsp-feedback/tsconfig.lib.json
@@ -13,10 +13,10 @@
     "src/**/*.spec.ts",
     "src/**/*.test.ts",
     "src/**/*.spec.js",
-    "src/**/*.test.js",
+    "src/**/*.test.js"
   ],
   "include": [
     "src/**/*.js",
-    "src/**/*.ts",
+    "src/**/*.ts"
   ]
 }
